* moving integration tests to own directory
